03_example_graphs.ipynb: Add further examples from FELINE paper
Add example of "a small DAG" from Figure 2 in the FELINE paper as
`small_DAG_FELINE()`, and "a didactic example of a DAG" (a tree DAG)
from Figure 6 in the FELINE paper as `tree_DAG()`.

NOTE: The `example_graph.py` was regenerated from the notebook.

This concludes the list of example small graphs (DAGs) from the FELINE
paper:
  Renê R. Veloso, Loïc Cerf, Wagner Meira Jr, Mohammed J. Zaki:
  "Reachability Queries in Very Large Graphs:
   A Fast Refined Online Search Approach",
  Proc. 17th International Conference on Extending Database Technology (EDBT),
  March 24-28, 2014, Athens, Greece: ISBN 978-3-89318065-3,
  http://openprocedings.org/EDBT/2014/paper_166.pdf