03_example_graphs.ipynb: Add figures from FELINE and DOI
Add the `images/feline/` directory with figures taken from the FELINE
paper, with and without their captions, and use the latter in
`03_example_graphs.ipynb` to show how example graphs could look like
according to the original paper.

While at it add a DOI (in the form of an URL) for the FELINE paper.