03_example_graphs.ipynb: Add graph from Fig. 9 in FELINE
Figure 9 in the FELINE paper is used to help explain the applicability
of level filter.  The example shows the case with a false-positive for
FELINE label, but which can be pruned in constant time with topological
levels - for query $r(h, g)$.

This graph is added as `levels_DAG_FELINE()` to the `example_graphs`
module.

While at it, the `matplotlib.pyplot` module is needed only for
demonstration purposes, so don't export it.