Add 09_git.ipynb notebook, with some exploration
Currently in this notebook we examine different ways of interacting with
Git repositories from Python, and do some explorations using the git.git
repository.

We examine and explore cloning the repository in an effective way,
generating the commit graph in NetworkX format, saving the graph,
computing and visualizing reachability labellings, and storing node
data.

Repositories are saved in (currently untracked) `repos/` subdirectory,
and play-datasets in various formats in the 'datasets/'.

Update the TODO list in index.ipynb (only fully done items count).