settings.ini: Add 'tables' as dependency for HDF5 output
PyTables is a package for managing hierarchical datasets and designed
to efficiently and easily cope with extremely large amounts of data.
PyTables is built on top of the HDF5 library, using the Python
language and the NumPy package.

PyTables module is used by Pandas to save DataFrames to HDF5 format.