09_git.ipynb: Fix error making CI to fail
The notebook didn't ensure that the `repos/` subdirectory existed by
creating it before trying to store git.git repository in it.

This made other commands to fail non-fatally (because they shell out),
and the error message occured much later in the notebook, when trying
to draw histogram of 'level' of nodes in the commit graph.

There was no problem in local installation, because 'repos/' directory
existed there.