09_git.ipynb: Make it work in Python 3.6, and other fixes
The `capture_output` named parameter was added to `subprocess.run` in
Python 3.7.  For GitHub CI (which uses Python 3.6) to work its use
had to be replaced by using `stdout=PIPE`.

While at it, emphasize where tests are in Markdown cells, and fix
outdated information about the repository used for tests.