Extract A.09_git_explore.ipynb out of 09_git.ipynb
Extracting the exploration part of the 09_git.ipynb notebook to an
appendix (to a separate A.09_git_explore.ipynb notebook) will make
managing and running it easier.

Note about the naming: the current naming convention for a per-notebook
appendix is A.<nn>_<name>_<purpose>.ipynb for <nn>_<name>.ipynb one.
This convention may change in the future.

The new notebook was added to table of contents in index.ipynb.