09_git.ipynb: Functions for saving graph and its data to a file
Add functions to save graph structure to a dataframe (as edge list), and
then saving it to a file and reading it from a file, with tests.

Add functions to compute reachability labels and store them in graph
object attributes, and to save them in a dataframe.  TODO: save
reachability labels and other per-node data to a file and read it from
a file.

TODO: add docstring for each new function.

_
Rename _repo_name_to_out_path() to _repo_path_to_out_path().

While at it, extend earlier tests a bit, and add information about
sources and sink nodes to the drawing of the commit graph for the
hellogitworld repository.

Fix imports.