09_git.ipynb: Split up _repo_name_to_out_basename()
Part 1 of refactoring of filename generation for storing the graph
structure of a commit graph (of a repository) and any directed graph,
and for storing per-node data like reachability labels.  The idea is to
have common code for this part of generating the commit graph and for
storing Dataframes with the data about graph.

The _repo_name_to_out_basename() got split into separate parts that are
responsible for separate parts of transformation.  The final function,
_savefile_name() generates whole name, not just the prefix / base of it.

The transformation now goes like this

   repo path -> repo name -> commit graph name -> save file name

Note however that the newly created _repo_graph_savefile() is not used
yet in commit_graph() function.