12_repos.ipynb: Don't rescan android-base repository
There is a problem with large repositories in which running git commands
might time out, like for example the 'android-base' repository.

The results of cloning the repositories, then extracting the commit
graph structure were added to the repository in the datasets/ directory
as *-commit_graph.adjlist.txt files.  Avoid re-creating those files.

This should fic the problem where number of nodes in 'android-base' is
computed to be 0, which results in division by zero when calculating
relative density of edges (number of edges to number of nodes).