Add 14_evaluation.ipynb notebook
This notebook is intended to compute and store the experimental results
of evaluation of how various reachability indices speed up reachability
queries.

It computes (or loads) the commit graph for the git.git repository and
computes backward topological levels and extended DFS intervals labeling
(or loads it).  It then selects a sample of node pairs (starting and
ending node), and checks whether those nodes are connected.

It also checks for true negatives and false positives for backward
topological levels (which can be used as a negative cut filter).

Currently there is no code (no functions, no classes) in the planned
module that is to correspond to this notebook.

.
Add to repository also *.csv.gz files storing graph structure in edge
list format, and computed per node data (as pandas DataFrames)

Updated table of contents in index.ipynb and README.md.