01_tools.ipynb: Add iGraph to the list of tools
iGraph is a library collection for creating and manipulating graphs and
analyzing networks. It is written in C and also exists as Python and
R packages.

Mentioned as one of possibilities in the Dmitry Zinoviev's talk titled
"Complex networks and their analysis in Python", filmed during the codebar
Literary Festival in collaboration with The Pragmatic Bookshelf
on Wednesday 23rd June 2021:
  https://www.youtube.com/watch?v=4SdYpv57fxQ