Upgrade directives with nbdev_migrate
Additionally, 03_example_graphs.ipynb had to be modified to not to
loose information, as originally the page description spanned two
lines (and nbdev_migrate lost the contents of the second line).

Also, _Reachability_labels_for_version_control_graphs.ipynb and
_Reachability_queries_in_large_graphs.ipynb were kept as is.  Those
files serve as original source to be put in nbdev-managed notebooks.